requirements.txt
streamlit
google-genai
streamlit-option-menu
pandas
